fix: portal.tsx 삭제
- portal.tsx 내부의 로직을 content.tsx로 이동
